update: move logic to activity
